git-svn-id: svn://svn.chromium.org/chrome/branches/ojan_experimental_branch/src@156 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/logging.h
@@ -480,8 +480,7 @@ class LogMessage {
 
   LogSeverity severity_;
   std::ostringstream stream_;
-  size_t message_start_;  // Offset of the start of the message (past prefix
-                          // info).
+  int message_start_;  // offset of the start of the message (past prefix info).
 
   DISALLOW_EVIL_CONSTRUCTORS(LogMessage);
 };
